Re-organizes internal Gradle plugins
