+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/features/FeatureMatrixTask.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.norris.gradle
+package io.dotanuki.norris.gradle.features
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
