+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/modules/models/ModuleConvention.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.norris.gradle.internal
+package io.dotanuki.norris.gradle.modules.models
 
 import org.gradle.api.GradleException
 import org.gradle.api.Project
