+++ gradle/plugins/src/main/kotlin/io/dotanuki/norris/gradle/modules/models/Versioning.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.norris.gradle.internal
+package io.dotanuki.norris.gradle.modules.models
 
 import java.text.SimpleDateFormat
 import java.util.Date
